--- lib/interactor/settings_interactor/settings_interactor.dart
@@ -1,7 +1,0 @@
-import 'package:places/interactor/settings_interactor/theme_state.dart';
-
-/// Интерактор настроек
-class SettingsInteractor {
-  // Текущая тема приложения
-  final ThemeState themeState = ThemeState();
-}
